Created new project with Python version 3.12
