+++ Code/Radio_Controller_Tank/.idea/libraries/MicroPython.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="MicroPython" type="python">
+    <CLASSES>
+      <root url="file://$APPLICATION_PLUGINS_DIR$/intellij-micropython/typehints/stdlib" />
+      <root url="file://$APPLICATION_PLUGINS_DIR$/intellij-micropython/typehints/micropython" />
+      <root url="file://$APPLICATION_PLUGINS_DIR$/intellij-micropython/typehints/rpi_pico" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
