+++ Code/Radio_Controller_Tank/.idea/misc.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Black">
+    <option name="sdkName" value="Python 3.12 (Radio_Controller_Tank)" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (Radio_Controller_Tank)" project-jdk-type="Python SDK" />
+</project>
